--- a/compiler/apps/playground/playwright.config.js
+++ b/compiler/apps/playground/playwright.config.js
@@ -23,7 +23,7 @@ export default defineConfig({
   // Test directory
   testDir: path.join(__dirname, '__tests__/e2e'),
   // If a test fails, retry it additional 2 times
-  retries: 2,
+  retries: 3,
   // Artifacts folder where screenshots, videos, and traces are stored.
   outputDir: 'test-results/',
   // Note: we only use text snapshots, so its safe to omit the host environment name