--- a/compiler/packages/babel-plugin-react-compiler/package.json
+++ b/compiler/packages/babel-plugin-react-compiler/package.json
@@ -9,7 +9,7 @@
     "!*.tsbuildinfo"
   ],
   "scripts": {
-    "build": "rimraf dist && scripts/build.js",
+    "build": "rimraf dist && tsup",
     "test": "./scripts/link-react-compiler-runtime.sh && yarn snap:ci",
     "jest": "yarn build && ts-node node_modules/.bin/jest",
     "snap": "node ../snap/dist/main.js",
@@ -17,7 +17,7 @@
     "snap:ci": "yarn snap:build && yarn snap",
     "ts:analyze-trace": "scripts/ts-analyze-trace.sh",
     "lint": "yarn eslint src",
-    "watch": "scripts/build.js --watch"
+    "watch": "yarn build --watch"
   },
   "dependencies": {
     "@babel/types": "^7.19.0"