--- a/compiler/packages/babel-plugin-react-compiler/tsconfig.json
+++ b/compiler/packages/babel-plugin-react-compiler/tsconfig.json
@@ -5,8 +5,6 @@
     "moduleResolution": "Bundler",
     "rootDir": "src",
     "outDir": "dist",
-    // https://github.com/microsoft/TypeScript/issues/30925
-    "tsBuildInfoFile": "dist/tsconfig.tsbuildinfo",
     "jsx": "react-jsxdev",
     // weaken strictness from preset
     "importsNotUsedAsValues": "remove",
@@ -16,7 +14,6 @@
     "target": "ES2015",
     // ideally turn off only during dev, or on a per-file basis
     "noUnusedLocals": false,
-    "composite": true,
     "removeComments": true
   },
   "exclude": [