--- a/compiler/packages/eslint-plugin-react-compiler/package.json
+++ b/compiler/packages/eslint-plugin-react-compiler/package.json
@@ -4,9 +4,9 @@
   "description": "ESLint plugin to display errors found by the React compiler.",
   "main": "dist/index.js",
   "scripts": {
-    "build": "rimraf dist && scripts/build.js",
-    "test": "tsc && jest",
-    "watch": "scripts/build.js --watch"
+    "build": "rimraf dist && tsup",
+    "test": "jest",
+    "watch": "yarn build --watch"
   },
   "files": [
     "dist"