--- a/compiler/packages/make-read-only-util/tsup.config.ts
+++ b/compiler/packages/make-read-only-util/tsup.config.ts
@@ -1,0 +1,29 @@
+import {defineConfig} from 'tsup';
+
+export default defineConfig({
+  entry: ['./src/makeReadOnly.ts'],
+  outDir: './dist',
+  splitting: false,
+  sourcemap: true,
+  dts: false,
+  bundle: true,
+  format: 'cjs',
+  platform: 'browser',
+  target: 'es2015',
+  banner: {
+    js: `/**
+ * Copyright (c) Meta Platforms, Inc. and affiliates.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ *
+ * @lightSyntaxTransform
+ * @noflow
+ * @nolint
+ * @preventMunge
+ * @preserve-invariant-messages
+ */
+
+"use no memo";`,
+  },
+});