--- a/compiler/packages/react-compiler-runtime/package.json
+++ b/compiler/packages/react-compiler-runtime/package.json
@@ -13,9 +13,9 @@
     "react": "^17.0.0 || ^18.0.0 || ^19.0.0 || ^0.0.0-experimental"
   },
   "scripts": {
-    "build": "rimraf dist && scripts/build.js",
+    "build": "rimraf dist && tsup",
     "test": "echo 'no tests'",
-    "watch": "scripts/build.js --watch"
+    "watch": "yarn build --watch"
   },
   "repository": {
     "type": "git",