--- a/compiler/scripts/release/shared/build-packages.js
+++ b/compiler/scripts/release/shared/build-packages.js
@@ -4,7 +4,7 @@ const {execHelper} = require('./utils');
 async function buildPackages(pkgNames) {
   const spinner = ora(`Building packages`).info();
   for (const pkgName of pkgNames) {
-    const command = `NODE_ENV=production yarn workspace ${pkgName} run build`;
+    const command = `NODE_ENV=production yarn workspace ${pkgName} run build --dts`;
     spinner.start(`Running: ${command}\n`);
     try {
       await execHelper(command);