… compiler/packages/babel-plugin-react-compiler/scripts/build.js
… compiler/packages/eslint-plugin-react-compiler/scripts/build.js
… compiler/packages/make-read-only-util/scripts/build.js
… compiler/packages/react-compiler-healthcheck/scripts/build.js
… compiler/yarn.lock